% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ada.classification-package.R
\docType{package}
\name{ada.classification-package}
\alias{ada.classification}
\alias{ada.classification-package}
\title{ada.classification: Classification Methods: K-means and Random Forests}
\description{
This package focuses on problems related to the classification of data into clusters or groups according to the characteristics of the data themselves. Toward this end, it contains functions for convenient implementation of k-means clustering and random forest decision trees with an example data set.
}
\author{
\strong{Maintainer}: John Duncan Hurt \email{hurt.john@utexas.edu}

Authors:
\itemize{
  \item Samantha Rabinowitz \email{samrabi@utexas.edu}
}

}
\keyword{internal}
