% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Pima Indians Diabetes Database}
\format{
A data frame with 768 rows and 9 columns:
\describe{
\item{Pregnancies}{the number of times the patient has been pregnant}
\item{Glucose}{the plasma glucose concentration measured for the patient after two hours in an oral glucose tolerance test (milligram/deciliter)}
\item{BloodPressure}{the diastolic blood pressure of the patient (mm Hg)}
\item{SkinThickness}{tricep skin fold thickness measured for the patient (mm)}
\item{Insulin}{2-Hour serum insulin level measured for the patient (mu U/ml)}
\item{BMI}{Body Mass Index of patient (weight in kg/(height in m)^2)}
\item{DiabetesPedigreeFunction}{score assigned based on family history of diabetes}
\item{Age}{age of the patient (years)}
\item{Outcome}{binary indicating diabetes diagnosis for patient: 1 for yes and 0 for no}
}
}
\source{
\url{https://www.kaggle.com/uciml/pima-indians-diabetes-database?select=diabetes.csv}
}
\usage{
diabetes
}
\description{
A dataset extracted from the National Institute of Diabetes and Digestive
and Kidney Diseases. The dataset contains a number of medically
relevant variables. It was created specifically (via extraction from a larger database)
to allow users to experiment with methods to predict the occurence of diabetes
in the patients. More information can be found in the URL provided below.
}
\keyword{datasets}
