% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansPlus.R
\name{kmeansPlus}
\alias{kmeansPlus}
\title{'kmeansPlus'}
\usage{
kmeansPlus(
  data,
  k,
  exclude = NULL,
  full.output = FALSE,
  plot = FALSE,
  x = NULL,
  y = NULL,
  iter.max = 10,
  nstart = 1,
  algorithm = c("Hartigan-Wong", "Lloyd", "Forgy", "MacQueen"),
  trace = FALSE
)
}
\arguments{
\item{data}{A matrix or dataframe containing numeric data in tidy format. This is equivalent to the 'x' argument from kmeans().}

\item{k}{Either the desired number of clusters, or a set of initial (distinct) cluster centers. This is equivalent to the 'centers' argument from kmeans().}

\item{exclude}{A vector, either numeric or character, indicating either indices or colnames of variables in 'data' that the user wishes to exclude. The excluded variable will not be removed from the tibble printed in the final output, but it will be excluded from all clustering and vizualisation procedures.}

\item{full.output}{A logical vector indicating whether the output of kmeansPlus() should include the full output of kmeans().}

\item{plot}{A logical vector indicating whether the output of kmeansPlus() should include a plot.}

\item{x}{A string (character vector of length 1) containing the colname of a variable from `data` to be plotted on the x axis (if plot = TRUE).}

\item{y}{A string (character vector of length 1) containing the colname of a variable from `data` to be plotted on the y axis (if plot = TRUE).}

\item{iter.max}{This is equivalent to the 'iter.max' argument from kmeans(), and is set to the same default.}

\item{nstart}{This is equivalent to the 'nstart' argument from kmeans(), and is set to the same default.}

\item{algorithm}{This is equivalent to the 'algorithm' argument from kmeans(), and is set to the same default.}

\item{trace}{This is equivalent to the 'trace' argument from kmeans(), and is set to the same default.}
}
\description{
A tool to aid in the implementation of conventional k means clustering, kmeansPlus() provides the normal functionality of stats::kmeans() with some modest additions. In short, it allows the user to quickly plot results from kmeans(), as well as add cluster identifiers to the input dataframe as a new variable.
}
\examples{
kmeansPlus(data = iris, k = 3, exclude = "Species", full.output = TRUE, plot = TRUE, x = "Sepal.Length", y = "Sepal.Width")
}
\keyword{classification,}
\keyword{k}
\keyword{means}
