% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rF_testing.R
\name{rF_testing}
\alias{rF_testing}
\title{A random forest classifier training and testing function}
\usage{
rF_testing(dat, s, x, model)
}
\arguments{
\item{dat}{dataset to perform test on}

\item{s}{proportion of dataset to be assigned to the 'training' dataset}

\item{x}{the classifying variable in the dataset}

\item{model}{written in the style of a regression, defines the classifying variable and the predictors the user wants to use to perform the test}
}
\description{
This function allows one to perform random forest classification (see \code{\link[randomForest]{randomForest}} for more information regarding the function to perform this classification method) of varying models on a training and testing set of the original data, the size of which is defined by the user.
}
\examples{
rF_testing(dat = iris, s = 0.5, x = Species,
model = Species ~ Petal.Length + Petal.Width + Sepal.Length)
}
\keyword{classification,}
\keyword{forest}
\keyword{random}
