% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ada.classification-package.R
\docType{package}
\name{ada.classification-package}
\alias{ada.classification}
\alias{ada.classification-package}
\title{ada.classification: Classification Methods: K-means and Random Forests}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: John Duncan Hurt \email{hurt.john@utexas.edu}

Authors:
\itemize{
  \item Samantha Rabinowitz \email{samrabi@utexas.edu}
}

}
\keyword{internal}
